feat(shared): add BilanType types
