+++ shared/src/types/BilanTypeSection.ts
@@ -1,0 +1,13 @@
+import type { BilanType } from './BilanType';
+
+export interface BilanTypeSection {
+  id: string;
+  bilanTypeId: string;
+  sectionId: string;
+  sortOrder: number;
+  settings?: unknown;
+  bilanType?: BilanType;
+  section?: unknown;
+}
+
+export type BilanTypeSections = BilanTypeSection[];
